refactor: 拆出usage_tracker
